works with the new stuff

+++ cmd/ghost_army_generator/ghost_army_generator.go
@@ -19,6 +19,7 @@ import (
  * August 18th: https://drive.google.com/file/d/1on1dfWUvGTxF_DDUKsyr1WJosiJrRfGO/view?usp=sharing
  * August 19th: https://drive.google.com/file/d/1AnIspF8WaZYQ4o8BSn7kb5DDZrK6E2Yh/view?usp=sharing
  * August 20th: https://drive.google.com/file/d/1z9crn3410-1_o3tJOnIswNxwj2rInVtJ/view?usp=sharing
+ * August 27th: https://drive.google.com/file/d/1y4zOJXXx_9KQD8pz_l5b46lNKrK-pkGN/view
  */
 
 type sortableEntries []ghostarmy.Entry
@@ -293,7 +294,7 @@ func main() {
 
 			if line[i] == "" {
 				// get random isp
-				entry.ISP = isps[entry.Userhash%int64(len(isps))]
+				entry.ISP = isps[uint64(entry.Userhash)%uint64(len(isps))]
 			} else {
 				// use actual isp
 				entry.ISP = line[i]
